[av] [LINUXDAR-5060] Replace numbers with strings to make log message more understandable.
Tidy types, and ensure Plugin tests aren't depending on the cpp file.

+++ av/modules/pluginimpl/StringUtils.cpp
@@ -106,7 +106,7 @@ long pluginimpl::getThreatStatus()
     {
         try
         {
-            long health = static_cast<long long>(j["threatHealth"]);
+            long health = static_cast<long>(j["threatHealth"]);
             return health;
         }
         catch (std::exception& exception)
